Changes for reboot

+++ C++/src/type/TypeManager.cpp
@@ -6,14 +6,14 @@ namespace Empire {
 std::string TypeData::ToString() {
 	std::stringstream ss;
 	ss << "Type Data ";
-	if (offset == UINT64_MAX) {
+	if (offset == UINT32_MAX) {
 		ss << "Class";
 	} else {
 		ss << "Field";
 	}
 	ss << " name " << name;
 	ss << ", size: " << size;
-	if (offset == UINT64_MAX) {
+	if (offset == UINT32_MAX) {
 		ss << ", fields: " << std::endl;
 		for (TypeData& data : fields) {
 			ss << "\t" << data.name << ", offset: " << data.offset << ", size: " << data.size << std::endl;
